# openapi.json
{"openapi":"3.0.2","info":{"title":"NER Service","version":"1.0"},"servers":[{"url":"/v1_0"}],"paths":{"/ping":{"get":{"summary":"Pong","operationId":"pong_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/requests/":{"get":{"tags":["requests"],"summary":"Read Requests","description":"Gets all requests in database paginated by skip and limit.\n\n- :param skip: number of request objects to skip\n- :param limit: max number of request objects to return\n- :return: list of requests","operationId":"read_requests_requests__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Requests Requests  Get","type":"array","items":{"$ref":"#/components/schemas/Request"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["requests"],"summary":"Create a request","description":"Create a request to perform Named Entity Recognition.\n- :param request: Request object containing path\n- :return: request object","operationId":"create_request_requests__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Request"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{req_id}":{"get":{"tags":["requests"],"summary":"Read Request","description":"Get request object by id.\n\n- :param req_id: ID of request\n- :return: request object","operationId":"read_request_requests__req_id__get","parameters":[{"required":true,"schema":{"title":"Req Id","type":"integer"},"name":"req_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Request"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requests/{req_id}/sentences":{"get":{"tags":["requests"],"summary":"Read Request Sentences","description":"Get all sentences in request of request id paginated by skip and limit.\n\n- :param req_id: ID of request\n- :param skip: number of sentence objects to skip\n- :param limit: max number of sentence objects to return\n- :return: list of sentences","operationId":"read_request_sentences_requests__req_id__sentences_get","parameters":[{"required":true,"schema":{"title":"Req Id","type":"integer"},"name":"req_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Request Sentences Requests  Req Id  Sentences Get","type":"array","items":{"$ref":"#/components/schemas/Sentence"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entities":{"get":{"tags":["entities"],"summary":"Read Entities","description":"Gets all entities in database paginated by skip and limit.\n\n- :param skip: number of entity objects to skip\n- :param limit: max number of entity objects to return\n- :param db: injected database\n- :return: list of entities","operationId":"read_entities_entities_get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Entities Entities Get","type":"array","items":{"$ref":"#/components/schemas/Entity"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entities/{entity_id}":{"get":{"tags":["entities"],"summary":"Get Entity","description":"Get entity object by id.\n\n- :param entity_id: ID of selected entity\n- :param db: injected database\n- :return: list of sentences","operationId":"get_entity_entities__entity_id__get","parameters":[{"required":true,"schema":{"title":"Entity Id","type":"integer"},"name":"entity_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Entity"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/entities/{entity_id}/sentences":{"get":{"tags":["entities"],"summary":"Search Entity Sentences","description":"Search for sentences by entity object by id.\n\n- :param entity_id: ID of selected entity\n- :param skip: number of sentence objects to skip\n- :param limit: max number of sentence objects to return\n- :return: list of sentences","operationId":"search_entity_sentences_entities__entity_id__sentences_get","parameters":[{"required":true,"schema":{"title":"Entity Id","type":"integer"},"name":"entity_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sentences":{"post":{"tags":["sentences"],"summary":"Get Entity Sentences","description":"Search for sentences by entity object by name.\n\n- :param entity_name: ID of selected entity\n- :param skip: number of sentence objects to skip\n- :param limit: max number of sentence objects to return\n- :return: list of sentences","operationId":"get_entity_sentences_sentences_post","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntitySearch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Entity":{"title":"Entity","required":["name","ent_type","id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"ent_type":{"title":"Ent Type","type":"string"},"id":{"title":"Id","type":"integer"}}},"EntitySearch":{"title":"EntitySearch","required":["entity_name"],"type":"object","properties":{"entity_name":{"title":"Entity Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Request":{"title":"Request","required":["path","id","status"],"type":"object","properties":{"path":{"title":"Path","type":"string"},"id":{"title":"Id","type":"integer"},"status":{"title":"Status","type":"string"},"sentences":{"title":"Sentences","type":"array","items":{"$ref":"#/components/schemas/Sentence"},"default":[]}}},"RequestCreate":{"title":"RequestCreate","required":["path"],"type":"object","properties":{"path":{"title":"Path","type":"string"}}},"Sentence":{"title":"Sentence","required":["text","id","req_id"],"type":"object","properties":{"text":{"title":"Text","type":"string"},"id":{"title":"Id","type":"integer"},"req_id":{"title":"Req Id","type":"integer"},"entities":{"title":"Entities","type":"array","items":{"$ref":"#/components/schemas/Entity"},"default":[]}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}